Extract institution search from monolithic repo
