+++ financial-connections/src/main/java/com/stripe/android/financialconnections/di/FinancialConnectionsSheetComponent.kt
@@ -13,7 +13,7 @@ import javax.inject.Singleton
 @Singleton
 @Component(
     modules = [
-        FinancialConnectionsSheetModule::class,
+        FinancialConnectionsSheetSharedModule::class,
         CoroutineContextModule::class,
         LoggingModule::class
     ]
