+++ financial-connections/src/main/java/com/stripe/android/financialconnections/di/FinancialConnectionsSheetNativeModule.kt
@@ -1,13 +1,17 @@
 package com.stripe.android.financialconnections.di
 
+import com.stripe.android.core.networking.ApiRequest
 import com.stripe.android.financialconnections.features.consent.ConsentSubcomponent
 import com.stripe.android.financialconnections.features.institutionpicker.InstitutionPickerSubcomponent
 import com.stripe.android.financialconnections.navigation.NavigationManager
+import com.stripe.android.financialconnections.network.FinancialConnectionsRequestExecutor
+import com.stripe.android.financialconnections.repository.FinancialConnectionsInstitutionsRepository
 import dagger.Module
 import dagger.Provides
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
+import javax.inject.Named
 import javax.inject.Singleton
 
 @Module(
@@ -22,5 +26,15 @@ internal class FinancialConnectionsSheetNativeModule {
     @Provides
     fun providesNavigationManager() = NavigationManager(
         CoroutineScope(SupervisorJob() + Dispatchers.Default)
+    )
+
+    @Singleton
+    @Provides
+    fun providesFinancialConnectionsInstitutionsRepository(
+        @Named(PUBLISHABLE_KEY) publishableKey: String,
+        requestExecutor: FinancialConnectionsRequestExecutor,
+        apiRequestFactory: ApiRequest.Factory
+    ) = FinancialConnectionsInstitutionsRepository(
+        publishableKey, requestExecutor, apiRequestFactory
     )
 }
