+++ financial-connections/src/main/java/com/stripe/android/financialconnections/domain/SearchInstitutions.kt
@@ -1,6 +1,7 @@
 package com.stripe.android.financialconnections.domain
 
 import com.stripe.android.financialconnections.model.InstitutionResponse
+import com.stripe.android.financialconnections.repository.FinancialConnectionsInstitutionsRepository
 import com.stripe.android.financialconnections.repository.FinancialConnectionsRepository
 import javax.inject.Inject
 
@@ -10,7 +11,7 @@ import javax.inject.Inject
  * If no input provided, this usecase returns featured institutions.
  */
 internal class SearchInstitutions @Inject constructor(
-    private val repository: FinancialConnectionsRepository
+    private val repository: FinancialConnectionsInstitutionsRepository
 ) {
     suspend operator fun invoke(
         clientSecret: String,
