+++ financial-connections/src/main/java/com/stripe/android/financialconnections/network/networkConstants.kt
@@ -1,0 +1,6 @@
+package com.stripe.android.financialconnections.network
+
+internal object NetworkConstants {
+    internal const val PARAMS_CLIENT_SECRET = "client_secret"
+    internal const val PARAMS_APPLICATION_ID = "application_id"
+}
