+++ financial-connections/src/main/java/com/stripe/android/financialconnections/repository/FinancialConnectionsRepositoryImpl.kt
@@ -1,45 +1,24 @@
 package com.stripe.android.financialconnections.repository
 
-import androidx.annotation.VisibleForTesting
-import com.stripe.android.core.exception.APIConnectionException
-import com.stripe.android.core.exception.APIException
-import com.stripe.android.core.exception.AuthenticationException
-import com.stripe.android.core.exception.InvalidRequestException
-import com.stripe.android.core.exception.PermissionException
-import com.stripe.android.core.exception.RateLimitException
-import com.stripe.android.core.model.parsers.StripeErrorJsonParser
 import com.stripe.android.core.networking.ApiRequest
-import com.stripe.android.core.networking.HTTP_TOO_MANY_REQUESTS
-import com.stripe.android.core.networking.StripeNetworkClient
-import com.stripe.android.core.networking.StripeRequest
-import com.stripe.android.core.networking.StripeResponse
-import com.stripe.android.core.networking.responseJson
+import com.stripe.android.core.networking.ApiRequest.Companion.API_HOST
 import com.stripe.android.financialconnections.di.PUBLISHABLE_KEY
 import com.stripe.android.financialconnections.model.FinancialConnectionsAccountList
 import com.stripe.android.financialconnections.model.FinancialConnectionsSession
 import com.stripe.android.financialconnections.model.FinancialConnectionsSessionManifest
 import com.stripe.android.financialconnections.model.FinancialConnectionsSessionManifest.FinancialConnectionsAuthorizationSession
 import com.stripe.android.financialconnections.model.GetFinancialConnectionsAcccountsParams
-import com.stripe.android.financialconnections.model.InstitutionResponse
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.json.Json
-import java.net.HttpURLConnection
+import com.stripe.android.financialconnections.network.FinancialConnectionsRequestExecutor
+import com.stripe.android.financialconnections.network.NetworkConstants.PARAMS_APPLICATION_ID
+import com.stripe.android.financialconnections.network.NetworkConstants.PARAMS_CLIENT_SECRET
 import javax.inject.Inject
 import javax.inject.Named
 
-internal class FinancialConnectionsApiRepository @Inject constructor(
+internal class FinancialConnectionsRepositoryImpl @Inject constructor(
     @Named(PUBLISHABLE_KEY) publishableKey: String,
-    private val stripeNetworkClient: StripeNetworkClient,
+    private val requestExecutor: FinancialConnectionsRequestExecutor,
     private val apiRequestFactory: ApiRequest.Factory
 ) : FinancialConnectionsRepository {
-
-    @VisibleForTesting
-    internal val json: Json = Json {
-        coerceInputValues = true
-        ignoreUnknownKeys = true
-        isLenient = true
-        encodeDefaults = true
-    }
 
     private val options = ApiRequest.Options(
         apiKey = publishableKey
@@ -53,7 +32,7 @@ internal class FinancialConnectionsApiRepository @Inject constructor(
             options = options,
             params = getFinancialConnectionsAcccountsParams.toParamMap()
         )
-        return executeRequest(
+        return requestExecutor.execute(
             financialConnectionsRequest,
             FinancialConnectionsAccountList.serializer()
         )
@@ -69,7 +48,10 @@ internal class FinancialConnectionsApiRepository @Inject constructor(
                 PARAMS_CLIENT_SECRET to clientSecret
             )
         )
-        return executeRequest(financialConnectionsRequest, FinancialConnectionsSession.serializer())
+        return requestExecutor.execute(
+            financialConnectionsRequest,
+            FinancialConnectionsSession.serializer()
+        )
     }
 
     override suspend fun generateFinancialConnectionsSessionManifest(
@@ -86,7 +68,7 @@ internal class FinancialConnectionsApiRepository @Inject constructor(
                 PARAMS_APPLICATION_ID to applicationId
             )
         )
-        return executeRequest(
+        return requestExecutor.execute(
             financialConnectionsRequest,
             FinancialConnectionsSessionManifest.serializer()
         )
@@ -102,7 +84,7 @@ internal class FinancialConnectionsApiRepository @Inject constructor(
                 PARAMS_CLIENT_SECRET to clientSecret
             )
         )
-        return executeRequest(
+        return requestExecutor.execute(
             financialConnectionsRequest,
             FinancialConnectionsSessionManifest.serializer()
         )
@@ -118,43 +100,9 @@ internal class FinancialConnectionsApiRepository @Inject constructor(
                 PARAMS_CLIENT_SECRET to clientSecret
             )
         )
-        return executeRequest(
+        return requestExecutor.execute(
             financialConnectionsRequest,
             FinancialConnectionsSessionManifest.serializer()
-        )
-    }
-
-    override suspend fun featuredInstitutions(clientSecret: String): InstitutionResponse {
-        val request = apiRequestFactory.createGet(
-            url = featuredInstitutionsUrl,
-            options = options,
-            params = mapOf(
-                PARAMS_CLIENT_SECRET to clientSecret,
-                "limit" to SEARCH_INSTITUTIONS_LIMIT
-            )
-        )
-        return executeRequest(
-            request,
-            InstitutionResponse.serializer()
-        )
-    }
-
-    override suspend fun searchInstitutions(
-        clientSecret: String,
-        query: String
-    ): InstitutionResponse {
-        val request = apiRequestFactory.createGet(
-            url = institutionsUrl,
-            options = options,
-            params = mapOf(
-                PARAMS_CLIENT_SECRET to clientSecret,
-                "query" to query,
-                "limit" to SEARCH_INSTITUTIONS_LIMIT
-            )
-        )
-        return executeRequest(
-            request,
-            InstitutionResponse.serializer()
         )
     }
 
@@ -171,68 +119,13 @@ internal class FinancialConnectionsApiRepository @Inject constructor(
                 "institution" to institutionId
             )
         )
-        return executeRequest(
+        return requestExecutor.execute(
             request,
             FinancialConnectionsAuthorizationSession.serializer()
         )
     }
 
-    private suspend fun <Response> executeRequest(
-        request: StripeRequest,
-        responseSerializer: KSerializer<Response>
-    ): Response = runCatching {
-        stripeNetworkClient.executeRequest(
-            request
-        )
-    }.fold(
-        onSuccess = { response ->
-            if (response.isError) {
-                throw handleApiError(response)
-            } else {
-                json.decodeFromString(
-                    responseSerializer,
-                    requireNotNull(response.body)
-                )
-            }
-        },
-        onFailure = {
-            throw APIConnectionException(
-                "Failed to execute $request",
-                cause = it
-            )
-        }
-    )
-
-    @Throws(
-        InvalidRequestException::class,
-        AuthenticationException::class,
-        APIException::class
-    )
-    private fun handleApiError(response: StripeResponse<String>): Exception {
-        val requestId = response.requestId?.value
-        val responseCode = response.code
-        val stripeError = StripeErrorJsonParser().parse(response.responseJson())
-        throw when (responseCode) {
-            HttpURLConnection.HTTP_BAD_REQUEST,
-            HttpURLConnection.HTTP_NOT_FOUND -> InvalidRequestException(
-                stripeError,
-                requestId,
-                responseCode
-            )
-            HttpURLConnection.HTTP_UNAUTHORIZED -> AuthenticationException(stripeError, requestId)
-            HttpURLConnection.HTTP_FORBIDDEN -> PermissionException(stripeError, requestId)
-            HTTP_TOO_MANY_REQUESTS -> RateLimitException(stripeError, requestId)
-            else -> APIException(stripeError, requestId, responseCode)
-        }
-    }
-
     internal companion object {
-        private const val API_HOST = "https://api.stripe.com"
-
-        private const val SEARCH_INSTITUTIONS_LIMIT = 8
-
-        internal const val PARAMS_CLIENT_SECRET = "client_secret"
-        internal const val PARAMS_APPLICATION_ID = "application_id"
         internal const val PARAMS_FULLSCREEN = "fullscreen"
         internal const val PARAMS_HIDE_CLOSE_BUTTON = "hide_close_button"
