+++ financial-connections/src/test/java/com/stripe/android/financialconnections/repository/FinancialConnectionsRepositoryImplTest.kt
@@ -7,7 +7,9 @@ import com.stripe.android.core.networking.StripeResponse
 import com.stripe.android.financialconnections.ApiKeyFixtures
 import com.stripe.android.financialconnections.model.BankAccount
 import com.stripe.android.financialconnections.model.FinancialConnectionsAccount
+import com.stripe.android.financialconnections.network.FinancialConnectionsRequestExecutor
 import com.stripe.android.financialconnections.test.readResourceAsString
+import com.stripe.android.financialconnections.utils.testJson
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
@@ -17,14 +19,17 @@ import org.mockito.kotlin.whenever
 import java.net.HttpURLConnection
 
 @ExperimentalCoroutinesApi
-class FinancialConnectionsApiRepositoryTest {
+class FinancialConnectionsRepositoryImplTest {
 
     private val mockStripeNetworkClient = mock<StripeNetworkClient>()
     private val apiRequestFactory = mock<ApiRequest.Factory>()
 
-    private val financialConnectionsApiRepository = FinancialConnectionsApiRepository(
+    private val financialConnectionsRepositoryImpl = FinancialConnectionsRepositoryImpl(
         publishableKey = ApiKeyFixtures.DEFAULT_PUBLISHABLE_KEY,
-        stripeNetworkClient = mockStripeNetworkClient,
+        requestExecutor = FinancialConnectionsRequestExecutor(
+            stripeNetworkClient = mockStripeNetworkClient,
+            json = testJson()
+        ),
         apiRequestFactory = apiRequestFactory
     )
 
@@ -37,7 +42,7 @@ class FinancialConnectionsApiRepositoryTest {
                 )
             )
 
-            val result = financialConnectionsApiRepository.getFinancialConnectionsSession("client_secret")
+            val result = financialConnectionsRepositoryImpl.getFinancialConnectionsSession("client_secret")
 
             assertThat(result.accounts.data.size).isEqualTo(1)
         }
@@ -51,7 +56,7 @@ class FinancialConnectionsApiRepositoryTest {
                 )
             )
 
-            val result = financialConnectionsApiRepository.getFinancialConnectionsSession("client_secret")
+            val result = financialConnectionsRepositoryImpl.getFinancialConnectionsSession("client_secret")
 
             assertThat(result.accounts.data.size).isEqualTo(1)
         }
@@ -65,7 +70,7 @@ class FinancialConnectionsApiRepositoryTest {
                 )
             )
 
-            val result = financialConnectionsApiRepository.getFinancialConnectionsSession("client_secret")
+            val result = financialConnectionsRepositoryImpl.getFinancialConnectionsSession("client_secret")
 
             assertThat(result.paymentAccount).isInstanceOf(FinancialConnectionsAccount::class.java)
         }
@@ -79,7 +84,7 @@ class FinancialConnectionsApiRepositoryTest {
                 )
             )
 
-            val result = financialConnectionsApiRepository.getFinancialConnectionsSession("client_secret")
+            val result = financialConnectionsRepositoryImpl.getFinancialConnectionsSession("client_secret")
 
             assertThat(result.paymentAccount).isInstanceOf(FinancialConnectionsAccount::class.java)
         }
@@ -93,7 +98,7 @@ class FinancialConnectionsApiRepositoryTest {
                 )
             )
 
-            val result = financialConnectionsApiRepository.getFinancialConnectionsSession("client_secret")
+            val result = financialConnectionsRepositoryImpl.getFinancialConnectionsSession("client_secret")
 
             assertThat(result.paymentAccount).isInstanceOf(FinancialConnectionsAccount::class.java)
         }
@@ -107,7 +112,7 @@ class FinancialConnectionsApiRepositoryTest {
                 )
             )
 
-            val result = financialConnectionsApiRepository.getFinancialConnectionsSession("client_secret")
+            val result = financialConnectionsRepositoryImpl.getFinancialConnectionsSession("client_secret")
 
             assertThat(result.paymentAccount).isInstanceOf(BankAccount::class.java)
         }
